backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==4.2.15
django-filter==24.3
django-stubs==5.0.4
django-stubs-ext==5.0.4
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
exceptiongroup==1.2.2
idna==3.8
iniconfig==2.0.0
mypy==1.11.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.2
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tomli==2.0.1
types-PyYAML==6.0.12.20240808
typing_extensions==4.12.2
urllib3==2.2.2
